requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-bootstrap5==23.4
idna==3.7
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.1
python3-openid==3.2.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==2.0.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.7.0
